UI/WinFrigg/Python/requirements.txt
numpy
pillow
scipy
matplotlib